requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==1.1.0
plotly==5.17.0
pandas==2.1.1